refactored last main integration test over to new Server format
