+++ tests/main/test_integration.py
@@ -1,12 +1,4 @@
 # -*- coding: utf-8 -*-
-from mantarray_desktop_app import CALIBRATION_NEEDED_STATE
-from mantarray_desktop_app import get_api_endpoint
-from mantarray_desktop_app import get_server_port_number
-from mantarray_desktop_app import system_state_eventually_equals
-from mantarray_desktop_app import wait_for_subprocesses_to_start
-import pytest
-import requests
-from stdlib_utils import confirm_port_in_use
 import base64
 import datetime
 import json
@@ -56,10 +48,11 @@ from mantarray_desktop_app import REFERENCE_VOLTAGE
 from mantarray_desktop_app import REFERENCE_VOLTAGE_UUID
 from mantarray_desktop_app import ROUND_ROBIN_PERIOD
 from mantarray_desktop_app import RunningFIFOSimulator
+from mantarray_desktop_app import server
 from mantarray_desktop_app import SERVER_READY_STATE
 from mantarray_desktop_app import SLEEP_FIRMWARE_VERSION_UUID
 from mantarray_desktop_app import SOFTWARE_RELEASE_VERSION_UUID
-from mantarray_desktop_app import START_RECORDING_TIME_INDEX_UUID,server
+from mantarray_desktop_app import START_RECORDING_TIME_INDEX_UUID
 from mantarray_desktop_app import system_state_eventually_equals
 from mantarray_desktop_app import SYSTEM_STATUS_UUIDS
 from mantarray_desktop_app import TIMESTEP_CONVERSION_FACTOR
@@ -92,25 +85,15 @@ from stdlib_utils import confirm_port_available
 from ..fixtures import fixture_fully_running_app_from_main_entrypoint
 from ..fixtures import fixture_patched_firmware_folder
 from ..fixtures import fixture_patched_shared_values_dict
-from ..fixtures import fixture_patched_short_calibration_script
-from ..fixtures import fixture_patched_start_recording_shared_dict
-from ..fixtures import fixture_patched_test_xem_scripts_folder,GENERIC_MAIN_LAUNCH_TIMEOUT_SECONDS 
 from ..fixtures import fixture_patched_xem_scripts_folder
-
-from ..fixtures import fixture_test_process_manager
-from ..fixtures import fixture_fully_running_app_from_main_entrypoint
-from ..fixtures import fixture_patched_firmware_folder
-from ..fixtures import fixture_patched_shared_values_dict
-from ..fixtures import fixture_patched_xem_scripts_folder
-
 from ..fixtures import fixture_test_process_manager
 from ..fixtures import fixture_test_process_manager_without_created_processes
+from ..fixtures import GENERIC_MAIN_LAUNCH_TIMEOUT_SECONDS
 from ..fixtures_file_writer import GENERIC_START_RECORDING_COMMAND
 from ..helpers import is_queue_eventually_empty
 
 __fixtures__ = [
     fixture_fully_running_app_from_main_entrypoint,
-    
     fixture_test_process_manager,
     fixture_patched_shared_values_dict,
     fixture_patched_xem_scripts_folder,
@@ -122,21 +105,20 @@ CALIBRATED_WAIT_TIME = 10
 STOP_MANAGED_ACQUISITION_WAIT_TIME = 40
 INTEGRATION_TEST_TIMEOUT = 720
 
+
 @pytest.mark.timeout(GENERIC_MAIN_LAUNCH_TIMEOUT_SECONDS)
 @pytest.mark.slow
 def test_send_xem_scripts_command__gets_processed_in_fully_running_app(
-    fully_running_app_from_main_entrypoint,
-    patched_xem_scripts_folder,
-    
+    fully_running_app_from_main_entrypoint, patched_xem_scripts_folder,
 ):
     app_info = fully_running_app_from_main_entrypoint(["--skip-mantarray-boot-up"])
     wait_for_subprocesses_to_start()
-    
-    
 
-    test_process_manager = app_info['object_access_inside_main']['process_manager']
-    test_process_monitor = app_info['object_access_inside_main']['process_monitor']
-    shared_values_dict=app_info['object_access_inside_main']['values_to_share_to_server']
+    test_process_manager = app_info["object_access_inside_main"]["process_manager"]
+    test_process_monitor = app_info["object_access_inside_main"]["process_monitor"]
+    shared_values_dict = app_info["object_access_inside_main"][
+        "values_to_share_to_server"
+    ]
     monitor_error_queue = test_process_monitor.get_fatal_error_reporter()
 
     response = requests.get(
@@ -170,8 +152,6 @@ def test_send_xem_scripts_command__gets_processed_in_fully_running_app(
 def test_system_states_and_recording_files_with_file_directory_passed_in_cmd_line_args(
     patched_xem_scripts_folder,
     patched_firmware_folder,
-    
-    
     fully_running_app_from_main_entrypoint,
     mocker,
 ):
@@ -200,8 +180,8 @@ def test_system_states_and_recording_files_with_file_directory_passed_in_cmd_lin
         ]
         app_info = fully_running_app_from_main_entrypoint(command_line_args)
         wait_for_subprocesses_to_start()
-        test_process_manager=app_info['object_access_inside_main']['process_manager']
-        test_process_monitor=app_info['object_access_inside_main']['process_monitor']
+        test_process_manager = app_info["object_access_inside_main"]["process_manager"]
+        test_process_monitor = app_info["object_access_inside_main"]["process_monitor"]
         response = requests.get(f"{get_api_endpoint()}system_status")
         assert response.status_code == 200
         assert response.json()["ui_status_code"] == str(
@@ -286,9 +266,7 @@ def test_system_states_and_recording_files_with_file_directory_passed_in_cmd_lin
 
         test_process_manager.soft_stop_processes()
 
-        fw_process = (
-            test_process_manager.get_file_writer_process()
-        )
+        fw_process = test_process_manager.get_file_writer_process()
         fw_process.close_all_files()
         actual_set_of_files = set(
             os.listdir(
@@ -301,19 +279,18 @@ def test_system_states_and_recording_files_with_file_directory_passed_in_cmd_lin
 
         test_process_manager.hard_stop_and_join_processes()
 
+
 @pytest.mark.slow
 @pytest.mark.timeout(INTEGRATION_TEST_TIMEOUT)
 def test_managed_acquisition_can_be_stopped_and_restarted_with_simulator(
     patched_xem_scripts_folder,
     patched_firmware_folder,
-    
-    
     fully_running_app_from_main_entrypoint,
 ):
     app_info = fully_running_app_from_main_entrypoint()
     wait_for_subprocesses_to_start()
-    test_process_manager=app_info['object_access_inside_main']['process_manager']
-    
+    test_process_manager = app_info["object_access_inside_main"]["process_manager"]
+
     assert system_state_eventually_equals(CALIBRATION_NEEDED_STATE, 5) is True
 
     response = requests.get(f"{get_api_endpoint()}start_calibration")
@@ -367,7 +344,7 @@ def test_managed_acquisition_can_be_stopped_and_restarted_with_simulator(
 
     test_process_manager.hard_stop_and_join_processes()
 
-import logging,click
+
 @pytest.mark.slow
 @pytest.mark.timeout(INTEGRATION_TEST_TIMEOUT)
 @freeze_time(
@@ -378,7 +355,6 @@ import logging,click
 def test_system_states_and_recorded_metadata_with_update_to_file_writer_directory(
     patched_xem_scripts_folder,
     patched_firmware_folder,
-
     fully_running_app_from_main_entrypoint,
     mocker,
 ):
@@ -394,19 +370,8 @@ def test_system_states_and_recorded_metadata_with_update_to_file_writer_director
 
     app_info = fully_running_app_from_main_entrypoint(["--skip-mantarray-boot-up"])
     wait_for_subprocesses_to_start()
-    werkzeug_logger=logging.getLogger('werkzeug')    
-    werkzeug_logger.setLevel(logging.ERROR)
-    werkzeug_logger.disabled=True
-    #werkzeug_logger.handlers.clear()
-    def secho(text, file=None, nl=None, err=None, color=None, **styles):
-        pass
-    
-    def echo(text, file=None, nl=None, err=None, color=None, **styles):
-        pass
-    
-    click.echo = echo
-    click.secho = secho    
-    test_process_manager=app_info['object_access_inside_main']['process_manager']
+
+    test_process_manager = app_info["object_access_inside_main"]["process_manager"]
 
     assert system_state_eventually_equals(SERVER_READY_STATE, 5) is True
     # response = requests.get(f"{get_api_endpoint()}system_status")
@@ -422,7 +387,7 @@ def test_system_states_and_recorded_metadata_with_update_to_file_writer_director
         assert response.status_code == 200
         response = requests.get(f"{get_api_endpoint()}boot_up")
         assert response.status_code == 200
-        assert system_state_eventually_equals(INSTRUMENT_INITIALIZING_STATE, 5) is True        
+        assert system_state_eventually_equals(INSTRUMENT_INITIALIZING_STATE, 5) is True
         # response = requests.get(f"{get_api_endpoint()}system_status")
         # assert response.status_code == 200
         # assert response.json()["ui_status_code"] == str(
@@ -495,10 +460,9 @@ def test_system_states_and_recorded_metadata_with_update_to_file_writer_director
 
         test_process_manager.soft_stop_processes()
 
-        
         fw_process = test_process_manager.get_file_writer_process()
         fw_process.close_all_files()
-        
+
         actual_set_of_files = set(
             os.listdir(
                 os.path.join(
@@ -507,7 +471,7 @@ def test_system_states_and_recorded_metadata_with_update_to_file_writer_director
             )
         )
         assert len(actual_set_of_files) == 24
-        
+
         # Tanner (6/15/20): Processes must be joined to avoid h5 errors with reading files, so clearing queues manually here in order to join
         test_process_manager.hard_stop_and_join_processes()
 
@@ -627,3 +591,186 @@ def test_system_states_and_recorded_metadata_with_update_to_file_writer_director
                         == REFERENCE_SENSOR_SAMPLING_PERIOD
                         * MICROSECONDS_PER_CENTIMILLISECOND
                     )
+
+
+@pytest.mark.slow
+@pytest.mark.timeout(INTEGRATION_TEST_TIMEOUT)
+def test_full_datapath(
+    patched_xem_scripts_folder,
+    patched_firmware_folder,
+    fully_running_app_from_main_entrypoint,
+):
+    app_info = fully_running_app_from_main_entrypoint()
+    wait_for_subprocesses_to_start()
+    test_process_manager = app_info["object_access_inside_main"]["process_manager"]
+
+    assert system_state_eventually_equals(CALIBRATION_NEEDED_STATE, 5) is True
+
+    da_error_queue = (
+        test_process_manager.queue_container().get_data_analyzer_error_queue()
+    )
+    da_out = test_process_manager.queue_container().get_data_analyzer_data_out_queue()
+
+    response = requests.get(f"{get_api_endpoint()}start_calibration")
+    assert response.status_code == 200
+    assert (
+        system_state_eventually_equals(CALIBRATED_STATE, CALIBRATED_WAIT_TIME) is True
+    )
+
+    acquisition_request = f"{get_api_endpoint()}start_managed_acquisition"
+    response = requests.get(acquisition_request)
+    assert response.status_code == 200
+    assert (
+        system_state_eventually_equals(
+            LIVE_VIEW_ACTIVE_STATE, LIVE_VIEW_ACTIVE_WAIT_TIME
+        )
+        is True
+    )
+
+    response = requests.get(f"{get_api_endpoint()}stop_managed_acquisition")
+    assert response.status_code == 200
+    assert (
+        system_state_eventually_equals(
+            CALIBRATED_STATE, STOP_MANAGED_ACQUISITION_WAIT_TIME
+        )
+        is True
+    )
+    response = requests.get(f"{get_api_endpoint()}get_available_data")
+    assert response.status_code == 200
+    actual = json.loads(response.text)
+    waveform_data_points = actual["waveform_data"]["basic_data"]["waveform_data_points"]
+    response = requests.get(f"{get_api_endpoint()}get_available_data")
+    assert response.status_code == 200
+    response = requests.get(f"{get_api_endpoint()}get_available_data")
+    assert response.status_code == 204
+    assert is_queue_eventually_empty(da_out) is True
+
+    test_well_index = 0
+    x_values = np.array(
+        [
+            (ROUND_ROBIN_PERIOD * (i + 1) // TIMESTEP_CONVERSION_FACTOR)
+            for i in range(
+                math.ceil(
+                    DATA_ANALYZER_BUFFER_SIZE_CENTIMILLISECONDS / ROUND_ROBIN_PERIOD
+                )
+            )
+        ]
+    )
+    sawtooth_points = signal.sawtooth(
+        x_values / FIFO_READ_PRODUCER_SAWTOOTH_PERIOD, width=0.5
+    )
+    scaling_factor = (test_well_index + 1) / 24 * 6
+    data_amplitude = int(FIFO_READ_PRODUCER_WELL_AMPLITUDE * scaling_factor)
+    test_data = np.array(
+        (
+            x_values,
+            (
+                (FIFO_READ_PRODUCER_DATA_OFFSET + data_amplitude * sawtooth_points)
+                - RAW_TO_SIGNED_CONVERSION_VALUE
+            ),
+        ),
+        dtype=np.int32,
+    )
+    pt = PipelineTemplate(
+        noise_filter_uuid=BUTTERWORTH_LOWPASS_30_UUID,
+        tissue_sampling_period=ROUND_ROBIN_PERIOD,
+    )
+
+    pipeline = pt.create_pipeline()
+    pipeline.load_raw_gmr_data(test_data, np.zeros(test_data.shape))
+    expected_well_data = pipeline.get_compressed_displacement()
+    actual_well_0_y_data = waveform_data_points["0"]["y_data_points"]
+    np.testing.assert_almost_equal(
+        actual_well_0_y_data[0],
+        expected_well_data[1][0] * MILLIVOLTS_PER_VOLT,
+        decimal=4,
+    )
+    np.testing.assert_almost_equal(
+        actual_well_0_y_data[9],
+        expected_well_data[1][9] * MILLIVOLTS_PER_VOLT,
+        decimal=4,
+    )
+
+    test_process_manager.hard_stop_and_join_processes()
+    assert is_queue_eventually_empty(da_error_queue) is True
+
+
+@pytest.mark.slow
+@pytest.mark.timeout(INTEGRATION_TEST_TIMEOUT)
+def test_app_shutdown__in_worst_case_while_recording_is_running(
+    patched_xem_scripts_folder,
+    patched_firmware_folder,
+    fully_running_app_from_main_entrypoint,
+    mocker,
+):
+    spied_logger = mocker.spy(main.logger, "info")
+    spied_server_logger = mocker.spy(server.logger, "info")
+    app_info = fully_running_app_from_main_entrypoint()
+    wait_for_subprocesses_to_start()
+    test_process_manager = app_info["object_access_inside_main"]["process_manager"]
+
+    assert system_state_eventually_equals(CALIBRATION_NEEDED_STATE, 5) is True
+
+    okc_process = test_process_manager.get_instrument_process()
+    fw_process = test_process_manager.get_file_writer_process()
+    da_process = test_process_manager.get_data_analyzer_process()
+
+    with tempfile.TemporaryDirectory() as expected_recordings_dir:
+        response = requests.get(
+            f"{get_api_endpoint()}update_settings?customer_account_uuid=curi&recording_directory={expected_recordings_dir}"
+        )
+        assert response.status_code == 200
+
+        response = requests.get(f"{get_api_endpoint()}start_calibration")
+        assert response.status_code == 200
+        assert system_state_eventually_equals(CALIBRATING_STATE, 5) is True
+
+        assert (
+            system_state_eventually_equals(CALIBRATED_STATE, CALIBRATED_WAIT_TIME)
+            is True
+        )
+
+        response = requests.get(f"{get_api_endpoint()}start_managed_acquisition")
+        assert response.status_code == 200
+
+        assert system_state_eventually_equals(BUFFERING_STATE, 5) is True
+
+        assert (
+            system_state_eventually_equals(
+                LIVE_VIEW_ACTIVE_STATE, LIVE_VIEW_ACTIVE_WAIT_TIME
+            )
+            is True
+        )
+
+        expected_barcode = GENERIC_START_RECORDING_COMMAND[
+            "metadata_to_copy_onto_main_file_attributes"
+        ][PLATE_BARCODE_UUID]
+        response = requests.get(
+            f"{get_api_endpoint()}start_recording?barcode={expected_barcode}&is_hardware_test_recording=False"
+        )
+        assert response.status_code == 200
+
+        assert system_state_eventually_equals(RECORDING_STATE, 5) is True
+        # response = requests.get(f"{get_api_endpoint()}system_status")
+        # assert response.status_code == 200
+        # assert response.json()["ui_status_code"] == str(
+        #     SYSTEM_STATUS_UUIDS[RECORDING_STATE]
+        # )
+
+        time.sleep(3)  # Tanner (6/15/20): This allows data to be written to files
+
+        response = requests.get(f"{get_api_endpoint()}shutdown")
+        assert response.status_code == 200
+
+    confirm_port_available(get_server_port_number(), timeout=10)
+
+    spied_server_logger.assert_any_call("Flask server successfully shut down.")
+
+    # Eli (12/4/20): currently, Flask immediately shuts down and communicates up to ProcessMonitor to start shutting everything else down. So for now need to sleep a bit before attempting to confirm everything else is shut down
+    time.sleep(10)
+
+    spied_logger.assert_any_call("Program exiting")
+
+    assert okc_process.is_alive() is False
+    assert fw_process.is_alive() is False
+    assert da_process.is_alive() is False
